Add %trackobjects and free func for TagLib::MPEG::File
This prevents programs from running into a segmentation fault in the
following situation:

  file = TagLib::MPEG::File.new("file.mp3")
  tag = file.id3v2_tag
  file = nil  # -> GC file
  tag.title  # -> segfault

The problem is that in taglib, the tag (and its frames) are "owned" by
File, which means that they are also deleted when File is deleted. This
make sense in C++, but leads to a segfault when we're not careful. With
this change, an exception is raised instead.