Add beginnings of API documentation in docs/
Extracting the documentation from the SWIG-generated wrappers would be
possible, but it would not be very nice to edit or read (because not
everything that SWIG generates should appear in documentation).

Therefore, the "docs" directory will host the API documentation. It will
be necessary to recreate the exact same structure with dummy methods.