Fix exception accessing TagLib::ID3v2::Tag.frame_list
After adding a Frame with TagLib::ID3v2::Tag.add_frame an
ObjectPreviouslyDeleted exception is thrown when accessing that frame.
To fix this, SWIG_RubyRemoveTracking is called on the C++ frame object
after adding it with Tag.add_frame. This is achieved by adding a SWIG
typemap for TagLib::ID3v2::Frame *. (See Issue #8)