Fixes for cross compiled Windows binary gem
Apparently with the newer MinGW-w64 toolchain, libstdc++ and libgcc are
no longer statically linked by default. So we now use -static-libgcc
-static-libstdc++ when building TagLib.

That fixes the problem for libtag.dll. But the extension modules which
are dynamically linked to libtag.dll also need libgcc and libstdc++.
libtag.dll would already have the libraries, but the symbols are not
exported any more (which was also changed apparently) and so the
extension modules can't use them. To fix this, we now also link them
statically for the extensions.

Using one shared libstdc++ DLL would also be possible, but without
stripping it's too big.

All in all this increases the size of the -x86-mingw32.gem from 1.4 MB
to 2.5 MB. But hey, at least it works.