Unlink and untrack tag in free func of MPEG and FLAC File (#14)
This was not done before because of the erroneous assumption that tag()
returned the same object as e.g. ID3v2Tag(). But of course it does not,
it's a separate object which is the union of multiple tags.

Thus it also needs to be unlinked and untracked when freeing MPEG::File
or FLAC::File. Otherwise when a new C++ object happens to get the same
pointer address as an old one, it is looked up in the tracked objects
and finds the old Ruby object (because the tracking entry is still
there). Because the underlying C++ object of the old Ruby object has
been freed, madness ensues.