Wrap taglib-1.11 Ogg/Picture methods for backward compatibility
Starting with taglib-1.11, pictures of a XiphComment are not accessible as
standard text fields but through a dedicated list. For backward compatibility,
the following methods are extended to handle pictures as standard text fields again.

The XiphComment::field_list_map method will report a METADATA_BLOCK_PICTURE field
containing the pictures, in the METADATA_BLOCK_PICTURE format, base64-encoded.

The XiphComment::add_field method will add/replace pictures to the dedicated list.
The value is expected to have the format associated to the key (METADATA_BLOCK_PICTURE
or COVERART). If the format is COVERART, the picture is converted to the
METADATA_BLOCK_PICTURE format.

The XiphComment::remove_field method will clear or remove pictures from the dedicated
list. Since the field_list_map does not expose COVERART pictures, this method only
support pictures in the METADATA_BLOCK_PICTURE format.

Wrappers are generated using taglib-1.7.2 headers and swig 3.0.7.